refactor: Integrating the game controller.
